feat: 스타카토 상세 사진 Viewpager 적용
