+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/moment/detail/ViewpagePhotoViewHolder.kt
@@ -1,0 +1,11 @@
+package com.woowacourse.staccato.presentation.moment.detail
+
+import androidx.recyclerview.widget.RecyclerView
+import com.woowacourse.staccato.databinding.ItemViewpagePhotoBinding
+
+class ViewpagePhotoViewHolder(private val binding: ItemViewpagePhotoBinding) :
+    RecyclerView.ViewHolder(binding.root) {
+    fun bind(item: String) {
+        binding.photoUrl = item
+    }
+}
