+++ android/Staccato_AN/app/src/main/res/layout/fragment_memory.xml
@@ -41,6 +41,7 @@
             android:id="@+id/sv_memory"
             android:layout_width="0dp"
             android:layout_height="0dp"
+            android:overScrollMode="never"
             android:scrollbars="none"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
