--- android/Staccato_AN/app/src/main/res/layout/fragment_moment_detail.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:bind="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="viewModel"
-            type="com.woowacourse.staccato.presentation.moment.detail.MomentDetailViewModel" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="10dp">
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_photo_horizontal"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:foregroundGravity="left"
-            android:orientation="horizontal"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:listitem="@layout/item_photo_horizontal" />
-
-        <TextView
-            android:id="@+id/tv_title"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="18dp"
-            android:layout_marginTop="20dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:text="@{viewModel.momentDetail.placeName}"
-            android:textAppearance="@style/Typography.Title1"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/rv_photo_horizontal"
-            tools:text="장소 제목" />
-
-        <TextView
-            android:id="@+id/tv_detail"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:text="@{viewModel.momentDetail.address}"
-            android:textAppearance="@style/Typography.Body4"
-            app:layout_constraintStart_toStartOf="@id/tv_title"
-            app:layout_constraintTop_toBottomOf="@id/tv_title"
-            tools:text="상세 주소" />
-
-        <TextView
-            android:id="@+id/tv_visited_at"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="20dp"
-            android:textAppearance="@style/Typography.Body2"
-            app:layout_constraintStart_toStartOf="@id/tv_title"
-            app:layout_constraintTop_toBottomOf="@id/tv_detail"
-            bind:visitedAt="@{viewModel.momentDetail.visitedAt}"
-            tools:text="2024년 7월 15일에 처음 방문했어요!" />
-
-        <com.google.android.material.divider.MaterialDivider
-            android:id="@+id/divider"
-            style="@style/DividerStyle"
-            android:layout_marginTop="17dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_visited_at" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</layout>
